+++ app/src/main/java/io/github/freewulkanowy/ui/modules/login/LoginActivity.kt
@@ -18,12 +18,13 @@ import io.github.freewulkanowy.databinding.ActivityLoginBinding
 import io.github.freewulkanowy.ui.base.BaseActivity
 import io.github.freewulkanowy.ui.modules.login.advanced.LoginAdvancedFragment
 import io.github.freewulkanowy.ui.modules.login.form.LoginFormFragment
+import io.github.freewulkanowy.ui.modules.login.onboarding.LoginOnboardingWarningFragment
+import io.github.freewulkanowy.ui.modules.login.onboarding.LoginOnboardingWelcomeFragment
 import io.github.freewulkanowy.ui.modules.login.recover.LoginRecoverFragment
 import io.github.freewulkanowy.ui.modules.login.studentselect.LoginStudentSelectFragment
 import io.github.freewulkanowy.ui.modules.login.symbol.LoginSymbolFragment
 import io.github.freewulkanowy.ui.modules.main.MainActivity
 import io.github.freewulkanowy.ui.modules.notifications.NotificationsFragment
-import io.github.freewulkanowy.ui.modules.onboarding.OnboardingActivity
 import io.github.freewulkanowy.utils.AppInfo
 import io.github.freewulkanowy.utils.InAppUpdateHelper
 import javax.inject.Inject
@@ -54,17 +55,13 @@ class LoginActivity : BaseActivity<LoginPresenter, ActivityLoginBinding>(), Logi
         val prefs = PreferenceManager.getDefaultSharedPreferences(this)
         val hasCompletedOnboarding = prefs.getBoolean("completedOnboarding", false)
 
-        if (!hasCompletedOnboarding) {
-            val onboardingIntent = Intent(this, OnboardingActivity::class.java)
-            startActivity(onboardingIntent)
-            finish()
-        }
-
         presenter.onAttachView(this)
         inAppUpdateHelper.checkAndInstallUpdates()
 
-        if (savedInstanceState == null) {
-            openFragment(LoginFormFragment.newInstance(), clearBackStack = true)
+        if (!hasCompletedOnboarding) {
+            openFragment(LoginOnboardingWarningFragment.newInstance(), clearBackStack = true)
+        } else if (savedInstanceState == null) {
+            navigateToLoginForm()
         }
     }
 
@@ -82,6 +79,14 @@ class LoginActivity : BaseActivity<LoginPresenter, ActivityLoginBinding>(), Logi
 
     fun showActionBar(show: Boolean) {
         supportActionBar?.run { if (show) show() else hide() }
+    }
+
+    fun navigateToOnboardingWelcomeFragment() {
+        openFragment(LoginOnboardingWelcomeFragment.newInstance())
+    }
+
+    fun navigateToLoginForm() {
+        openFragment(LoginFormFragment.newInstance(), clearBackStack = true)
     }
 
     fun navigateToSymbolFragment(loginData: LoginData) {
