+++ app/src/main/java/io/github/freewulkanowy/ui/modules/login/onboarding/LoginOnboardingView.kt
@@ -1,0 +1,7 @@
+package io.github.freewulkanowy.ui.modules.login.onboarding
+
+import io.github.freewulkanowy.ui.base.BaseView
+
+interface LoginOnboardingView : BaseView {
+    fun initView()
+}
