--- app/src/main/java/io/github/freewulkanowy/ui/modules/onboarding/OnboardingView.kt
@@ -1,8 +1,0 @@
-package io.github.freewulkanowy.ui.modules.onboarding
-
-import io.github.freewulkanowy.ui.base.BaseView
-
-interface OnboardingView : BaseView {
-
-    fun initView()
-}
