+++ app/src/main/res/layout/fragment_onboarding_warning.xml
@@ -14,9 +14,8 @@
             android:id="@+id/onboardingIcon"
             android:layout_width="96dp"
             android:layout_height="96dp"
-            android:layout_marginTop="32dp"
             android:src="@mipmap/ic_launcher"
-            app:layout_constraintBottom_toTopOf="@id/onboardingText"
+            app:layout_constraintBottom_toTopOf="@id/onboarding1Text"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
@@ -24,42 +23,56 @@
             app:srcCompat="@drawable/ic_launcher_foreground_mono" />
 
         <TextView
-            android:id="@+id/onboardingText"
+            android:id="@+id/onboarding1Text"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="32dp"
             android:layout_marginTop="16dp"
             android:fontFamily="sans-serif-medium"
             android:gravity="center"
-            android:text="@string/onboarding_2_title"
+            android:text="@string/onboarding_1_title"
             android:textSize="28sp"
-            app:layout_constraintBottom_toTopOf="@id/onboardingDescription"
+            app:layout_constraintBottom_toTopOf="@id/onboarding1Description"
             app:layout_constraintTop_toBottomOf="@id/onboardingIcon" />
 
         <TextView
-            android:id="@+id/onboardingDescription"
+            android:id="@+id/onboarding1Description"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginHorizontal="32dp"
             android:layout_marginTop="16dp"
             android:gravity="center"
             android:textSize="16sp"
-            app:layout_constraintBottom_toTopOf="@id/onboardingCloseBtn"
-            app:layout_constraintTop_toBottomOf="@id/onboardingText"
+            app:layout_constraintBottom_toTopOf="@id/onboardingNextBtn"
+            app:layout_constraintTop_toBottomOf="@id/onboarding1Text"
             tools:maxLines="15"
-            android:text="@string/onboarding_2_message" />
+            android:text="@string/onboarding_1_message" />
 
         <com.google.android.material.button.MaterialButton
-            android:id="@+id/onboardingCloseBtn"
-            android:layout_width="match_parent"
+            android:id="@+id/onboardingBackBtn"
+            style="@style/Widget.Material3.Button.OutlinedButton"
+            android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginRight="32dp"
-            android:layout_marginLeft="32dp"
+            android:layout_marginTop="32dp"
+            android:layout_marginStart="32dp"
             android:layout_marginBottom="32dp"
-            android:text="@string/onboarding_close_btn"
-            android:textSize="16sp"
-            app:layout_constraintBottom_toBottomOf="parent"
+            android:text="@string/onboarding_back_btn"
+            app:icon="@drawable/ic_arrow_left"
+            app:iconGravity="start"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/onboardingNextBtn"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="32dp"
+            android:layout_marginEnd="32dp"
+            android:layout_marginBottom="32dp"
+            android:text="@string/onboarding_next_btn"
+            app:icon="@drawable/ic_arrow_right"
+            app:iconGravity="end"
             app:layout_constraintEnd_toEndOf="parent"
-            tools:ignore="MissingConstraints" />
+            app:layout_constraintBottom_toBottomOf="parent" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </ScrollView>
